refactor: mv vpool-xyk-amm to amm
