+++ packages/SwingSet/misc-tools/measure-metering/measure.js
@@ -50,7 +50,7 @@ async function run() {
   const targetFn = new URL('measurement-target.js', import.meta.url).pathname;
   const zoeFn = new URL('measurement-zoe.js', import.meta.url).pathname;
   const autoswapFn = new URL(
-    '../../../inter-protocol/src/vpool-xyk-amm/multipoolMarketMaker.js',
+    '../../../inter-protocol/src/amm/multipoolMarketMaker.js',
     import.meta.url,
   ).pathname;
   const autoswapBundle = await bundleSource(autoswapFn);
