+++ packages/inter-protocol/src/amm/constantProduct/AMM-Trade.jpeg
