+++ packages/inter-protocol/src/amm/constantProduct/invariants.js
