+++ packages/inter-protocol/src/exported.js
@@ -1,2 +1,2 @@
-import './vpool-xyk-amm/types.js';
+import './amm/types.js';
 import './vaultFactory/types.js';
