+++ packages/inter-protocol/test/amm/constantProduct/test-calcDeltaX.js
@@ -4,7 +4,7 @@
 import { test } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 import { AmountMath } from '@agoric/ertp';
 
-import { calcDeltaXSellingX } from '../../../src/vpool-xyk-amm/constantProduct/core.js';
+import { calcDeltaXSellingX } from '../../../src/amm/constantProduct/core.js';
 import { setupMintKits } from './setupMints.js';
 
 const doTest = (t, x, y, deltaY, expectedDeltaX) => {
