+++ packages/inter-protocol/test/test-interchainPool.js
@@ -6,7 +6,7 @@ import { E, Far } from '@endo/far';
 import { makePromiseKit } from '@endo/promise-kit';
 import { makeNodeBundleCache } from './bundleTool.js';
 import { setUpZoeForTest } from './supports.js';
-import { setupAmmServices } from './amm/vpool-xyk-amm/setup.js';
+import { setupAmmServices } from './amm/setup.js';
 import { startInterchainPool } from '../src/proposals/econ-behaviors.js';
 
 /** @template T @typedef {import('@endo/promise-kit').PromiseKit<T>} PromiseKit */
