+++ packages/inter-protocol/test/vaultFactory/test-vaultFactory.js
@@ -67,7 +67,7 @@ const contractRoots = {
   faucet: './test/vaultFactory/faucet.js',
   liquidate: './src/vaultFactory/liquidateMinimum.js',
   VaultFactory: './src/vaultFactory/vaultFactory.js',
-  amm: './src/vpool-xyk-amm/multipoolMarketMaker.js',
+  amm: './src/amm/multipoolMarketMaker.js',
   reserve: './src/reserve/assetReserve.js',
 };
